Support JSONPath in attribute external IDs (#6)
Define a new WithJSONPathAttributeNames JSON option.
When the WithJSONPathAttributeNames option is enabled, and an attribute or child entity's external ID starts with '$', the attribute or child entity is converted from the value(s) matched by the JSONPath in the original object.
If the attribute is not a list (i.e. it is configured without list set to true) and its external ID is a JSONPath:
If the JSONPath matches no value, then the attribute is set to null.
If the JSONPath matches a single value, then the attribute is set to that value.
If the JSONPath matches more than one value, then an error is returned.
If the attribute is a list (i.e. it is configured with list set to true) and its external ID is a JSONPath:
If the JSONPath matches no value, then the attribute is set to null.
If the JSONPath matches a list of one or more values, then the attribute is set to that list.
Document the WithComplexAttributeNameDelimiter option as deprecated since WithJSONPathAttributeNames is functionally a superset.