Add optional Logger interface to Adapter Server (#101)
This PR introduces a new optional parameter to set a logger when
creating a new adapter server.

If the logger is set, all adapter calls to `GetPage` will have their
context enriched with a child logger and related request fields (e.g.
datasource type, ID, etc.). It's up to the adapter implementation to
retrieve the logger from the context and use it.

Helper functions to add/get the logger from the context are added:
- `ContextWithLogger`
- `LoggerFromContext`

Additionally, add the optional `clientId/tenantId` fields to the Adapter
protobuf spec. This is useful for callers of `GetPage` if they want to
add additional metadata to their request.